Measure the decision path, then optimize what the measurements found
Implements milestone 1 of the deep optimization review, plus the two follow-on
passes its measurements justified. Ordered deliberately: instrument first, so
every later change is answerable with a number.

Instrumentation (app/telemetry.py)
  DistributionRegistry keeps bounded recent-sample latency/size distributions
  with p50/p95/p99, and EventLoopMonitor samples how late the loop runs a timer
  it should have run on time. The decision path is staged end to end
  (event.lock_wait, engine.prepare/canonicalize/native/materialize, ledger.write,
  model_lab.enqueue, account.mark/place, sse.snapshot), reported at
  /api/runtime under "performance". Metric names are code-level literals and the
  registry caps distinct names, so an event id can never become a metric label.

Native detachment (native_engine/src/lib.rs)
  evaluate_json now takes Python<'_> and runs parse/score/encode inside
  py.detach. The request is taken as an owned String rather than &str on
  purpose: a &str borrows the Python object's buffer, and reading it while
  detached means touching Python-owned memory with no interpreter reference
  held. Without detaching, offloading the call to a worker thread would move the
  work but not release the loop, because the worker would still hold the GIL.

Engine phase split (app/engine.py)
  SignalEngine.evaluate is now prepare_request -> evaluate_prepared ->
  materialize_signals, with evaluate() preserved as the sync wrapper and
  evaluate_async() offloading only the native round trip. Only the canonical
  request string crosses to the worker. PreparedRequest snapshots the lineage
  versions, closing a real latent bug: a calibration installed mid-evaluation
  could previously stamp a signal with lineage its own decision hash never
  covered. Measured on 4 concurrent prop-heavy events, p99 event-loop lag drops
  211ms -> 27ms (7.9x) and throughput rises 57 -> 133 evaluations/second.

Benchmarks (tools/bench_workload.py, tools/bench_decision.py, benchmarks/)
  Three checked-in immutable fixtures (small/normal/heavy), covering ingest,
  prepare, native, materialize, request/output bytes, persistence, and an A/B
  event-loop lag harness. Emits JSON; a non-blocking CI job uploads the report.

What the measurements then found, and what was done about it
  Native scoring is only 10-18% of a decision; `prepare` and ingestion were
  ~85%. Profiling those showed both were dominated by pure string normalization
  repeated on a tiny set of distinct values -- base_market_type alone ran ~13
  times per quote. Memoizing the single-argument identity functions (bounded
  lru_cache, hit rates on /api/runtime) cut ingest 3.3x and prepare 1.5x, which
  no amount of compiler tuning on the Rust side could have achieved.

Bounded write lanes (app/dbwriter.py)
  One dedicated worker thread and one bounded queue per store, replacing shared
  -executor to_thread calls. Ledger and account writes stay durable-before-
  continue; model-lab research rows are queued rather than awaited, leaving the
  per-event critical section. A full queue applies backpressure instead of
  dropping -- discarding model-lab observations would bias the calibration
  training sample -- and shutdown drains before the stores close.

Deliberately NOT done
  No [profile.release] tuning: thin LTO + codegen-units=1 measured ~1% on native
  p50, inside the run-to-run spread, for a 7s -> 44s build. panic="abort" is
  additionally unsafe here, since PyO3 converts panics into Python exceptions.
  Rationale recorded in Cargo.toml so it is re-measured, not re-litigated.

Also: Ruff expanded to B/ASYNC/UP/SIM/RUF/PERF with every ignore justified
against real findings, and the native engine import now fails closed at the
first scoring call instead of at import, so storage/security/identity tests can
run without the compiled extension. Production startup still asserts it.

Full results in docs/decision-path-performance.md.
Suite 764 passed / 1 skipped; ruff, mypy, cargo fmt/clippy/test clean.

+++ app/projections.py
@@ -148,7 +148,7 @@ def _poisson_partition(mu: float, line: float) -> DiscreteLineProbability:
     if mu > _EXACT_COUNT_MAX_MEAN:
         return _continuity_corrected(mu, sd, line, "poisson", {"mu": mu})
     log_mu = math.log(mu)
-    upper = max(int(math.ceil(mu + 12.0 * sd + 20.0)), math.floor(line) + 2)
+    upper = max(math.ceil(mu + 12.0 * sd + 20.0), math.floor(line) + 2)
     # Each pmf is computed independently in log space, so a tiny P(X=0) never
     # zeros the recurrence the way a forward `exp(-mu)` product would.
     pmf = [math.exp(-mu + k * log_mu - math.lgamma(k + 1)) for k in range(upper + 1)]
@@ -168,7 +168,7 @@ def _negbin_partition(mu: float, variance: float, line: float) -> DiscreteLinePr
     if mu > _EXACT_COUNT_MAX_MEAN:
         return _continuity_corrected(mu, sd, line, "negative_binomial", params)
     log_p, log_1mp = math.log(p), math.log1p(-p)
-    upper = max(int(math.ceil(mu + 12.0 * sd + 20.0)), math.floor(line) + 2)
+    upper = max(math.ceil(mu + 12.0 * sd + 20.0), math.floor(line) + 2)
     pmf = [
         math.exp(math.lgamma(k + r) - math.lgamma(r) - math.lgamma(k + 1)
                  + r * log_p + k * log_1mp)
